Improve many parsing-related functions to use size_t

+++ generic/tclCompCmdsSZ.c
@@ -1465,7 +1465,7 @@ void
 TclSubstCompile(
     Tcl_Interp *interp,
     const char *bytes,
-    int numBytes,
+    size_t numBytes,
     int flags,
     int line,
     CompileEnv *envPtr)
